% Auto-generated: do not edit by hand
\name{copyButton}

\alias{copyButton}

\title{CopyButton component}

\description{
CopyButton is an button that when clicked copies text to the clipboard.  It has an input value which is the content to be copied. Additional properties of typical Dash components
}

\usage{
copyButton(children=NULL, id=NULL, disabled=NULL, value=NULL,
n_clicks=NULL, n_clicks_timestamp=NULL, className=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. Array that holds components to render}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{disabled}{Character | logical. Indicates wheather the user can interact with the element}

\item{value}{Character. The value to copy.}

\item{n_clicks}{Numeric. Number of times the button has been clicked.}

\item{n_clicks_timestamp}{Numeric. Last time the button was clicked.}

\item{className}{Character. Often used with CSS to style elements with common properties.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

